fix(ci): hosted sync and line-limit checks - fixtures missed terminal-head delta git calls
